[FE] refactor: 디렉토리명 및 파일명 리팩토링
