+++ frontend/src/app/main/layout.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { AppSidebar } from '@/components/Sidebar';
+import { AppSidebar } from '@/components/sidebar';
 import { SidebarProvider } from '@/components/ui/sidebar';
 
 export default function MainLayout({
