+++ frontend/src/app/main/projects/annotation/[id]/page.tsx
@@ -10,7 +10,7 @@ type ProjectWithModel = Project & {
   modelType?: string;
 };
 
-const OSDViewer = dynamic(
+const AnnotationViewer = dynamic(
   () => import('@/components/annotation/annotation-viewer'),
   {
     ssr: false,
@@ -52,7 +52,10 @@ export default function Annotation({
     <div>
       <div>
         {/* OSDViewer에 어노테이션 데이터 대신 빈 배열과 modelType 전달 */}
-        <OSDViewer annotations={[]} modelType={project?.modelType || 'MULTI'} />
+        <AnnotationViewer
+          annotations={[]}
+          modelType={project?.modelType || 'MULTI'}
+        />
       </div>
     </div>
   );
