+++ frontend/src/components/sidebar/sidebar-bottom-menu/index.tsx
