+++ frontend/src/components/sidebar/sidebar-profile/index.tsx
